refactor the structure of code
