+++ src/modules/check/check.controller.js
@@ -1,4 +1,4 @@
-const config = require("../../../config");
+const config = require("../../../config/variables.config");
 const db = require("../../../core/database/db.router");
 const statusCode = require("../../../core/base/statusCode");
 const router = require("../../../core/base/enum/route.enum");
