+++ src/modules/user/user.helper.js
@@ -1,5 +1,5 @@
 const crypto = require("crypto");
-const config = require("../../../config");
+const config = require("../../../config/variables.config");
 
 class UserHelper {
   hashPassword(str) {
