+++ src/utils/JsonToObject.js
@@ -1,0 +1,10 @@
+const parceJsonToObject = (str) => {
+  try {
+    const obj = JSON.parse(str);
+    return obj;
+  } catch (e) {
+    return {};
+  }
+};
+
+module.exports = parceJsonToObject;
