+++ src/workers.js
@@ -1,9 +1,7 @@
 const http = require("http");
 const https = require("https");
-const fs = require("fs");
-const path = require("path");
 const url = require("url");
-// const _data = require("./data");
+const _data = require("../core/database/db.router");
 
 const workers = {};
